+++ it/server/src/test/java/com/walmartlabs/concord/it/server/ProjectFileIT.java
@@ -38,7 +38,8 @@ import java.util.*;
 import static com.walmartlabs.concord.it.common.ITUtils.archive;
 import static com.walmartlabs.concord.it.common.ServerClient.assertLog;
 import static com.walmartlabs.concord.it.common.ServerClient.waitForCompletion;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 public class ProjectFileIT extends AbstractServerIT {